Crate proj_sys

Low-level bindings for PROJ v7.1.x

This is a *-sys crate; you shouldn’t use its API directly. See the proj crate for general use.

A guide to the functions can be found here: https://proj.org/development/reference/functions.html.

By default, the crate will search for an existing libproj (via PROJ v7.1.x) installation on your system using pkg-config.

If an acceptable installation is not found, proj-sys will attempt to build libproj from source bundled in the crate.

Features

bundled_proj - forces building libproj from source even if an acceptable version could be found on your system. Note that SQLite3 and libtiff must be present on your system if you wish to use this feature, and that it builds libproj without its native network functionality; you will have to implement your own set of callbacks if you wish to make use of them (see the proj crate for an example).

Structs

P5_FACTORS

PJ_AREA

PJ_ELLPS

PJ_ENU

PJ_GEOD

PJ_GRID_INFO

PJ_INFO

PJ_INIT_INFO

PJ_LIST

PJ_LP

PJ_LPZ

PJ_LPZT

PJ_OBJ_LIST

PJ_OPERATION_FACTORY_CONTEXT

PJ_OPK

PJ_PRIME_MERIDIANS

PJ_PROJ_INFO

PJ_UNITS

PJ_UV

PJ_UVW

PJ_UVWT

PJ_XY

PJ_XYZ

PJ_XYZT

PJconsts

PROJ_CRS_INFO

\brief Structure given overall description of a CRS.

PROJ_CRS_LIST_PARAMETERS

\brief Structure describing optional parameters for proj_get_crs_list();

PROJ_FILE_API

File API callbacks

PROJ_FILE_HANDLE

PROJ_NETWORK_HANDLE

PROJ_UNIT_INFO

\brief Structure given description of a unit.

projCtx_t

Constants

PJ_CATEGORY_PJ_CATEGORY_COORDINATE_OPERATION

PJ_CATEGORY_PJ_CATEGORY_CRS

PJ_CATEGORY_PJ_CATEGORY_DATUM

PJ_CATEGORY_PJ_CATEGORY_ELLIPSOID

PJ_CATEGORY_PJ_CATEGORY_PRIME_MERIDIAN

PJ_COMPARISON_CRITERION_PJ_COMP_EQUIVALENT

The objects are equivalent for the purpose of coordinate operations. They can differ by the name of their objects, identifiers, other metadata. Parameters may be expressed in different units, provided that the value is (with some tolerance) the same once expressed in a common unit.

PJ_COMPARISON_CRITERION_PJ_COMP_EQUIVALENT_EXCEPT_AXIS_ORDER_GEOGCRS

Same as EQUIVALENT, relaxed with an exception that the axis order of the base CRS of a DerivedCRS/ProjectedCRS or the axis order of a GeographicCRS is ignored. Only to be used with DerivedCRS/ProjectedCRS/GeographicCRS

PJ_COMPARISON_CRITERION_PJ_COMP_STRICT

All properties are identical.

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_CARTESIAN

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_DATETIMETEMPORAL

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_ELLIPSOIDAL

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_ORDINAL

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_PARAMETRIC

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_SPHERICAL

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_TEMPORALCOUNT

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_TEMPORALMEASURE

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_UNKNOWN

PJ_COORDINATE_SYSTEM_TYPE_PJ_CS_TYPE_VERTICAL

PJ_DEFAULT_CTX

PJ_DIRECTION_PJ_FWD

PJ_DIRECTION_PJ_IDENT

PJ_DIRECTION_PJ_INV

PJ_GUESSED_WKT_DIALECT_PJ_GUESSED_NOT_WKT

Not WKT / unrecognized

PJ_GUESSED_WKT_DIALECT_PJ_GUESSED_WKT1_ESRI

ESRI variant of WKT1

PJ_GUESSED_WKT_DIALECT_PJ_GUESSED_WKT1_GDAL

\ref WKT1

PJ_GUESSED_WKT_DIALECT_PJ_GUESSED_WKT2_2015

\ref WKT2_2015

PJ_GUESSED_WKT_DIALECT_PJ_GUESSED_WKT2_2018

Deprecated alias for PJ_GUESSED_WKT2_2019

PJ_GUESSED_WKT_DIALECT_PJ_GUESSED_WKT2_2019

\ref WKT2_2019

PJ_LOG_LEVEL_PJ_LOG_DEBUG

PJ_LOG_LEVEL_PJ_LOG_DEBUG_MAJOR

PJ_LOG_LEVEL_PJ_LOG_DEBUG_MINOR

PJ_LOG_LEVEL_PJ_LOG_ERROR

PJ_LOG_LEVEL_PJ_LOG_NONE

PJ_LOG_LEVEL_PJ_LOG_TELL

PJ_LOG_LEVEL_PJ_LOG_TRACE

PJ_PROJ_STRING_TYPE_PJ_PROJ_4

cf osgeo::proj::io::PROJStringFormatter::Convention::PROJ_4

PJ_PROJ_STRING_TYPE_PJ_PROJ_5

cf osgeo::proj::io::PROJStringFormatter::Convention::PROJ_5

PJ_TYPE_PJ_TYPE_BOUND_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_COMPOUND_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_CONCATENATED_OPERATION

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_CONVERSION

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_CRS

Abstract type, not returned by proj_get_type()

PJ_TYPE_PJ_TYPE_DATUM_ENSEMBLE

PJ_TYPE_PJ_TYPE_DYNAMIC_GEODETIC_REFERENCE_FRAME

PJ_TYPE_PJ_TYPE_DYNAMIC_VERTICAL_REFERENCE_FRAME

PJ_TYPE_PJ_TYPE_ELLIPSOID

PJ_TYPE_PJ_TYPE_ENGINEERING_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_ENGINEERING_DATUM

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_GEOCENTRIC_CRS

Abstract type, not returned by proj_get_type()

PJ_TYPE_PJ_TYPE_GEODETIC_CRS

Abstract type, not returned by proj_get_type()

PJ_TYPE_PJ_TYPE_GEODETIC_REFERENCE_FRAME

PJ_TYPE_PJ_TYPE_GEOGRAPHIC_2D_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_GEOGRAPHIC_3D_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_GEOGRAPHIC_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_OTHER_COORDINATE_OPERATION

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_OTHER_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_PARAMETRIC_DATUM

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_PRIME_MERIDIAN

PJ_TYPE_PJ_TYPE_PROJECTED_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_TEMPORAL_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_TEMPORAL_DATUM

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_TRANSFORMATION

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_UNKNOWN

PJ_TYPE_PJ_TYPE_VERTICAL_CRS

proj_get_type() will never return that type, but PJ_TYPE_GEOGRAPHIC_2D_CRS or PJ_TYPE_GEOGRAPHIC_3D_CRS.

PJ_TYPE_PJ_TYPE_VERTICAL_REFERENCE_FRAME

PJ_WKT_TYPE_PJ_WKT1_ESRI

cf osgeo::proj::io::WKTFormatter::Convention::WKT1_ESRI

PJ_WKT_TYPE_PJ_WKT1_GDAL

cf osgeo::proj::io::WKTFormatter::Convention::WKT1_GDAL

PJ_WKT_TYPE_PJ_WKT2_2015

cf osgeo::proj::io::WKTFormatter::Convention::WKT2

PJ_WKT_TYPE_PJ_WKT2_2015_SIMPLIFIED

cf osgeo::proj::io::WKTFormatter::Convention::WKT2_SIMPLIFIED

PJ_WKT_TYPE_PJ_WKT2_2018

Deprecated alias for PJ_WKT2_2019

PJ_WKT_TYPE_PJ_WKT2_2018_SIMPLIFIED

Deprecated alias for PJ_WKT2_2019

PJ_WKT_TYPE_PJ_WKT2_2019

cf osgeo::proj::io::WKTFormatter::Convention::WKT2_2019

PJ_WKT_TYPE_PJ_WKT2_2019_SIMPLIFIED

cf osgeo::proj::io::WKTFormatter::Convention::WKT2_2019_SIMPLIFIED

PROJ_CRS_EXTENT_USE_PJ_CRS_EXTENT_BOTH

Test coordinate operation extent against both CRS extent.

PROJ_CRS_EXTENT_USE_PJ_CRS_EXTENT_INTERSECTION

Test coordinate operation extent against the intersection of both CRS extent.

PROJ_CRS_EXTENT_USE_PJ_CRS_EXTENT_NONE

Ignore CRS extent

PROJ_CRS_EXTENT_USE_PJ_CRS_EXTENT_SMALLEST

Test coordinate operation against the smallest of both CRS extent.

PROJ_GRID_AVAILABILITY_USE_PROJ_GRID_AVAILABILITY_DISCARD_OPERATION_IF_MISSING_GRID

Completely discard an operation if a required grid is missing.

PROJ_GRID_AVAILABILITY_USE_PROJ_GRID_AVAILABILITY_IGNORED

Ignore grid availability at all. Results will be presented as if all grids were available.

PROJ_GRID_AVAILABILITY_USE_PROJ_GRID_AVAILABILITY_KNOWN_AVAILABLE

Results will be presented as if grids known to PROJ (that is registered in the grid_alternatives table of its database) were available. Used typically when networking is enabled.

PROJ_GRID_AVAILABILITY_USE_PROJ_GRID_AVAILABILITY_USED_FOR_SORTING

Grid availability is only used for sorting results. Operations where some grids are missing will be sorted last.

PROJ_INTERMEDIATE_CRS_USE_PROJ_INTERMEDIATE_CRS_USE_ALWAYS

Always search for intermediate CRS.

PROJ_INTERMEDIATE_CRS_USE_PROJ_INTERMEDIATE_CRS_USE_IF_NO_DIRECT_TRANSFORMATION

Only attempt looking for intermediate CRS if there is no direct transformation available.

PROJ_INTERMEDIATE_CRS_USE_PROJ_INTERMEDIATE_CRS_USE_NEVER

Only attempt looking for intermediate CRS if there is no direct transformation available.

PROJ_OPEN_ACCESS_PROJ_OPEN_ACCESS_CREATE

Create access. File should be truncated to 0-byte if already existing. Equivalent to “w+b”

PROJ_OPEN_ACCESS_PROJ_OPEN_ACCESS_READ_ONLY

Read-only access. Equivalent to “rb”

PROJ_OPEN_ACCESS_PROJ_OPEN_ACCESS_READ_UPDATE

Read-update access. File should be created if not existing. Equivalent to “r+b”

PROJ_SPATIAL_CRITERION_PROJ_SPATIAL_CRITERION_PARTIAL_INTERSECTION

The area of validity of transforms should at least intersect the area of interest.

PROJ_SPATIAL_CRITERION_PROJ_SPATIAL_CRITERION_STRICT_CONTAINMENT

The area of validity of transforms should strictly contain the are of interest.

PROJ_VERSION_MAJOR

PROJ_VERSION_MINOR

PROJ_VERSION_PATCH

Statics

pj_release

Functions

proj_angular_input⚠

proj_angular_output⚠

proj_area_create⚠

proj_area_destroy⚠

proj_area_set_bbox⚠

proj_as_proj_string⚠

proj_as_projjson⚠

proj_as_wkt⚠

proj_assign_context⚠

@cond Doxygen_Suppress

proj_cleanup⚠

proj_clone⚠

proj_concatoperation_get_step⚠

proj_concatoperation_get_step_count⚠

proj_context_clone⚠

proj_context_create⚠

proj_context_destroy⚠

proj_context_errno⚠

proj_context_get_database_metadata⚠

proj_context_get_database_path⚠

proj_context_get_url_endpoint⚠

proj_context_get_use_proj4_init_rules⚠

proj_context_get_user_writable_directory⚠

proj_context_guess_wkt_dialect⚠

proj_context_is_network_enabled⚠

proj_context_set_autoclose_database⚠

proj_context_set_ca_bundle_path⚠

proj_context_set_database_path⚠

proj_context_set_enable_network⚠

proj_context_set_file_finder⚠

proj_context_set_fileapi⚠

proj_context_set_network_callbacks⚠

proj_context_set_search_paths⚠

proj_context_set_sqlite3_vfs_name⚠

proj_context_set_url_endpoint⚠

proj_context_use_proj4_init_rules⚠

proj_coord⚠

proj_coordoperation_create_inverse⚠

proj_coordoperation_get_accuracy⚠

proj_coordoperation_get_grid_used⚠

proj_coordoperation_get_grid_used_count⚠

proj_coordoperation_get_method_info⚠

proj_coordoperation_get_param⚠

proj_coordoperation_get_param_count⚠

proj_coordoperation_get_param_index⚠

proj_coordoperation_get_towgs84_values⚠

proj_coordoperation_has_ballpark_transformation⚠

proj_coordoperation_is_instantiable⚠

proj_create⚠

@cond Doxygen_Suppress

proj_create_argv⚠

proj_create_crs_to_crs⚠

proj_create_crs_to_crs_from_pj⚠

proj_create_from_database⚠

proj_create_from_name⚠

proj_create_from_wkt⚠

proj_create_operation_factory_context⚠

@endcond

proj_create_operations⚠

proj_crs_get_coordinate_system⚠

proj_crs_get_coordoperation⚠

proj_crs_get_datum⚠

proj_crs_get_datum_ensemble⚠

proj_crs_get_datum_forced⚠

proj_crs_get_geodetic_crs⚠

proj_crs_get_horizontal_datum⚠

proj_crs_get_sub_crs⚠

proj_crs_info_list_destroy⚠

proj_cs_get_axis_count⚠

proj_cs_get_axis_info⚠

proj_cs_get_type⚠

proj_datum_ensemble_get_accuracy⚠

proj_datum_ensemble_get_member⚠

proj_datum_ensemble_get_member_count⚠

proj_degree_input⚠

proj_degree_output⚠

proj_destroy⚠

proj_dmstor⚠

proj_download_file⚠

proj_dynamic_datum_get_frame_reference_epoch⚠

proj_ellipsoid_get_parameters⚠

proj_errno⚠

proj_errno_reset⚠

proj_errno_restore⚠

proj_errno_set⚠

proj_errno_string⚠

proj_factors⚠

proj_geod⚠

proj_get_area_of_use⚠

proj_get_authorities_from_database⚠

proj_get_codes_from_database⚠

proj_get_crs_info_list_from_database⚠

proj_get_crs_list_parameters_create⚠

proj_get_crs_list_parameters_destroy⚠

proj_get_ellipsoid⚠

proj_get_id_auth_name⚠

proj_get_id_code⚠

proj_get_name⚠

proj_get_non_deprecated⚠

proj_get_prime_meridian⚠

proj_get_remarks⚠

proj_get_scope⚠

proj_get_source_crs⚠

proj_get_suggested_operation⚠

proj_get_target_crs⚠

proj_get_type⚠

proj_get_units_from_database⚠

proj_grid_cache_clear⚠

proj_grid_cache_set_enable⚠

proj_grid_cache_set_filename⚠

proj_grid_cache_set_max_size⚠

proj_grid_cache_set_ttl⚠

proj_grid_get_info_from_database⚠

proj_grid_info⚠

proj_identify⚠

proj_info⚠

proj_init_info⚠

proj_int_list_destroy⚠

proj_is_crs⚠

proj_is_deprecated⚠

proj_is_download_needed⚠

proj_is_equivalent_to⚠

proj_is_equivalent_to_with_ctx⚠

proj_list_angular_units⚠

proj_list_destroy⚠

proj_list_ellps⚠

proj_list_get⚠

proj_list_get_count⚠

proj_list_operations⚠

proj_list_prime_meridians⚠

proj_list_units⚠

proj_log_func⚠

proj_log_level⚠

proj_lp_dist⚠

proj_lpz_dist⚠

proj_normalize_for_visualization⚠

@endcond Doxygen_Suppress

proj_operation_factory_context_destroy⚠

proj_operation_factory_context_set_allow_ballpark_transformations⚠

proj_operation_factory_context_set_allow_use_intermediate_crs⚠

proj_operation_factory_context_set_allowed_intermediate_crs⚠

proj_operation_factory_context_set_area_of_interest⚠

proj_operation_factory_context_set_crs_extent_use⚠

proj_operation_factory_context_set_desired_accuracy⚠

proj_operation_factory_context_set_discard_superseded⚠

proj_operation_factory_context_set_grid_availability_use⚠

proj_operation_factory_context_set_spatial_criterion⚠

proj_operation_factory_context_set_use_proj_alternative_grid_names⚠

proj_pj_info⚠

proj_prime_meridian_get_parameters⚠

proj_roundtrip⚠

proj_rtodms⚠

proj_string_list_destroy⚠

@endcond

proj_todeg⚠

proj_torad⚠

proj_trans⚠

proj_trans_array⚠

proj_trans_generic⚠

proj_unit_list_destroy⚠

proj_uom_get_info_from_database⚠

proj_xy_dist⚠

proj_xyz_dist⚠

Type Definitions

PJ

PJ_CATEGORY

\brief Object category.

PJ_COMPARISON_CRITERION

Comparison criterion.

PJ_CONTEXT

PJ_COORDINATE_SYSTEM_TYPE

Type of coordinate system.

PJ_DIRECTION

PJ_FACTORS

PJ_GUESSED_WKT_DIALECT

\brief Guessed WKT “dialect”.

PJ_LOG_FUNCTION

PJ_LOG_LEVEL

PJ_OPERATIONS

PJ_PROJ_STRING_TYPE

\brief PROJ string version.

PJ_TYPE

\brief Object type.

PJ_WKT_TYPE

\brief WKT version.

PROJ_CRS_EXTENT_USE

Specify how source and target CRS extent should be used to restrict candidate operations (only taken into account if no explicit area of interest is specified.

PROJ_GRID_AVAILABILITY_USE

Describe how grid availability is used.

PROJ_INTERMEDIATE_CRS_USE

Describe if and how intermediate CRS should be used

PROJ_OPEN_ACCESS

Open access / mode

PROJ_SPATIAL_CRITERION

Spatial criterion to restrict candidate operations.

PROJ_STRING_LIST

\brief Type representing a NULL terminated list of NULL-terminate strings.

proj_file_finder

Callback to resolve a filename to a full path

proj_network_close_cbk_type

Network access: close callback

proj_network_get_header_value_cbk_type

Network access: get HTTP headers

proj_network_open_cbk_type

Network access: open callback

proj_network_read_range_type

Network access: read range

wchar_t

Unions

PJ_COORD